multi-agent-demo/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langfuse>=2.0.0
prometheus-client>=0.21.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.5.0
boto3>=1.35.0
